Include OpenSSL installation
Zabbix API will not respond unless there is a Certificate that is valid or not, (curl is ok with insecure connections in function GetZabbixAuthToken) and installing OpenSSL will trigger the script to generate self signed certs. API will work after. Always a good idea to change these certs later with "real" ones or Let's Encrypt.